requirements.txt
PyMuPDF==1.23.14
Pillow==10.1.0
numpy>=2.3.0
